corona/corona/requirements.txt
Django==1.6.5
pyOpenSSL==0.13
rq==0.3.9
django-rq==0.6.1
apns-client==0.2
redis==2.10.1
django-redis==3.6.2
pymongo==2.7.1
requests==2.3.0
django-rq-dashboard==0.2.1
djangorestframework==2.3.14
django-cors-headers>=0.10
hiredis>=0.1.4
mongoengine
django-websocket-redis
wsaccel
uwsgi
rq-scheduler
blinker
django-redisboard